:recycle: Remove ImagePickEventAdaptor
