+++ Sources/Domain/SpotPost/Components/SpotPostImage.swift
@@ -4,16 +4,41 @@ import SwiftUI
 struct SpotPostImage: View {
     @Environment(\.geocoder) var geocoder
 
-    @State var showsActionSheet: Bool = false
+    @State var isPresentingPhotoLibrary: Bool = false
+    @State var presentingAlertType: AlertType?
     @State var error: Error?
 
     @Binding var photoLibraryResult: PhotoLibraryResult?
     @Binding var photoLibraryPlacemark: Placemark?
 
+    enum AlertType: Int, Identifiable {
+        case openSetting
+        case choseNoPermission
+
+        var id: Self { self }
+    }
+
     var body: some View {
         Button (
             action: {
-                showsActionSheet = true
+                switch photoLibrary.prepareActionType() {
+                case nil:
+                    isPresentingPhotoLibrary = true
+                case .openSettingApp:
+                    presentingAlertType = .openSetting
+                case .requestAuthorization:
+                    Task {
+                        let status = await photoLibrary.requestAuthorization()
+                        switch status {
+                        case .authorized, .limited:
+                            isPresentingPhotoLibrary = true
+                        case .denied, .restricted, .notDetermined:
+                            presentingAlertType = .choseNoPermission
+                        @unknown default:
+                            assertionFailure("New case \(status)")
+                        }
+                    }
+                }
             },
             label: {
                 if let image = photoLibraryResult?.image {
@@ -41,21 +66,52 @@ struct SpotPostImage: View {
                 }
             })
             .buttonStyle(PlainButtonStyle())
-            .adaptImagePickEvent(
-                showsActionSheet: $showsActionSheet,
-                photoLibraryResult: Binding(get: {
-                    photoLibraryResult
-                }, set: { photoLibraryResult in
-                    Task {
-                        if let photoLibraryResultLocation = photoLibraryResult?.location {
-                            photoLibraryPlacemark = try? await geocoder.reverseGeocode(location: photoLibraryResultLocation).first
-                        }
-                        self.photoLibraryResult = photoLibraryResult
-                    }
-                }),
-                error: $error
+            .sheet(
+                isPresented: $isPresentingPhotoLibrary,
+                content: {
+                    PhotoLibraryView(
+                        photoLibrary: photoLibrary,
+                        photoLibraryResult: Binding(get: {
+                            photoLibraryResult
+                        }, set: { photoLibraryResult in
+                            Task {
+                                if let photoLibraryResultLocation = photoLibraryResult?.location {
+                                    photoLibraryPlacemark = try? await geocoder.reverseGeocode(location: photoLibraryResultLocation).first
+                                }
+                                self.photoLibraryResult = photoLibraryResult
+                            }
+                        }),
+                        error: $error
+                    )
+                }
             )
+            .alert(item: $presentingAlertType, content: { alertType in
+                switch alertType {
+                case .openSetting:
+                    return Alert(
+                        title: Text("画像を選択できません"),
+                        message: Text("フォトライブラリのアクセスが許可されていません。設定アプリから許可をしてください"),
+                        primaryButton: .default(Text("設定を開く"), action: openSetting),
+                        secondaryButton: .cancel()
+                    )
+                case .choseNoPermission:
+                    return Alert(
+                        title: Text("アクセスを拒否しました"),
+                        message: Text("フォトライブラリのアクセスが拒否されました。操作を続ける場合は設定アプリから許可をしてください"),
+                        primaryButton: .default(Text("設定を開く"), action: openSetting),
+                        secondaryButton: .cancel()
+                    )
+                }
+            })
             .handle(error: $error)
+    }
+
+
+    private func openSetting() {
+        let settingURL = URL(string: UIApplication.openSettingsURLString)!
+        if UIApplication.shared.canOpenURL(settingURL) {
+            UIApplication.shared.open(settingURL)
+        }
     }
 }
